Updates for lci tactical plugin ROS2
