+++ carma/launch/guidance.launch
@@ -193,9 +193,6 @@
     <!-- Stop Controlled Intersection Tactical Plugin -->
     <include file="$(find stop_controlled_intersection_tactical_plugin)/launch/stop_controlled_intersection_tactical_plugin.launch" />
 
-    <!-- Light Controlled Intersection Tactical Plugin -->
-    <include file="$(find light_controlled_intersection_tactical_plugin)/launch/light_controlled_intersection_tactical_plugin.launch" />
-
   </group>
 
   <!-- Trajectory Plan Speed Visualizer -->
