+++ light_controlled_intersection_tactical_plugin/include/light_controlled_intersection_tactical_plugin/light_controlled_intersection_tactical_plugin_node.hpp
@@ -18,31 +18,10 @@
 
 #include <rclcpp/rclcpp.hpp>
 #include <functional>
-#include <boost/uuid/uuid_generators.hpp>
-#include <boost/uuid/uuid_io.hpp>
-#include <boost/geometry.hpp>
-#include <boost/shared_ptr.hpp>
-#include <lanelet2_core/geometry/Point.h>
-#include <trajectory_utils/trajectory_utils.hpp>
-#include <trajectory_utils/conversions/conversions.hpp>
-#include <Eigen/Core>
-#include <Eigen/Geometry>
-#include <Eigen/LU>
-#include <Eigen/SVD>
-#include <lanelet2_core/primitives/Lanelet.h>
-#include <lanelet2_core/geometry/LineString.h>
-#include <carma_ros2_utils/containers/containers.hpp>
-#include <carma_planning_msgs/msg/trajectory_plan_point.hpp>
-#include <carma_planning_msgs/msg/trajectory_plan.hpp>
-#include <carma_planning_msgs/srv/plan_trajectory.hpp>
-#include <carma_planning_msgs/msg/maneuver.hpp>
-#include <autoware_msgs/msg/lane.hpp>
-#include <carma_debug_ros2_msgs/msg/trajectory_curvature_speeds.hpp>
-#include <basic_autonomy_ros2/basic_autonomy.hpp>
-#include <basic_autonomy_ros2/helper_functions.hpp>
 
 #include <carma_guidance_plugins/tactical_plugin.hpp>
 #include "light_controlled_intersection_tactical_plugin/light_controlled_intersection_tactical_plugin_config.hpp"
+#include "light_controlled_intersection_tactical_plugin/light_controlled_intersection_tactical_plugin.hpp"
 
 /**
  * \brief Macro definition to enable easier access to fields shared across the maneuver types
@@ -59,35 +38,6 @@
 
 namespace light_controlled_intersection_tactical_plugin
 {
-  using PointSpeedPair = basic_autonomy::waypoint_generation::PointSpeedPair;
-  using GeneralTrajConfig = basic_autonomy::waypoint_generation::GeneralTrajConfig;
-  using DetailedTrajConfig = basic_autonomy::waypoint_generation::DetailedTrajConfig;
-
-  enum TSCase {
-    CASE_1 = 1,
-    CASE_2 = 2,
-    CASE_3 = 3,
-    CASE_4 = 4,
-    CASE_5 = 5,
-    CASE_6 = 6,
-    CASE_7 = 7,
-    CASE_8 = 8,
-  };
-
-  struct TrajectoryParams {
-    double a1_ = 0;
-    double v1_ = 0;
-    double x1_ = 0;
-
-    double a2_ = 0;
-    double v2_ = 0;
-    double x2_ = 0;
-
-    double a3_ = 0;
-    double v3_ = 0;
-    double x3_ = 0;
-  };
-
   /**
    * \brief ROS node for the LightControlledIntersectionTransitPluginNode
    * 
@@ -98,40 +48,7 @@ namespace light_controlled_intersection_tactical_plugin
     // LightControlledIntersectionTransitPluginNode configuration
     Config config_;
 
-    // CARMA Streets Variables
-    // timestamp for msg received from carma streets
-    uint32_t street_msg_timestamp_ = 0.0;
-    // scheduled stop time
-    uint32_t scheduled_stop_time_ = 0.0;
-    // scheduled enter time
-    uint32_t scheduled_enter_time_ = 0.0;
-    // scheduled depart time
-    uint32_t scheduled_depart_time_ = 0.0;
-    // scheduled latest depart time
-    uint32_t scheduled_latest_depart_time_ = 0.0;
-    // flag to show if the vehicle is allowed in intersection
-    bool is_allowed_int_ = false;
-
-    double speed_limit_ = 11.176; // Approximate speed limit; 25 mph by default
-    boost::optional<TSCase> last_case_;
-    boost::optional<bool> is_last_case_successful_;
-    carma_planning_msgs::msg::TrajectoryPlan last_trajectory_;
-
-    carma_planning_msgs::msg::VehicleState ending_state_before_buffer_; //state before applying extra points for curvature calculation that are removed later
-
-    double epsilon_ = 0.001; //Small constant to compare (double) 0.0 with
-
-    // downtrack of host vehicle
-    double current_downtrack_ = 0.0;
-
-    double last_successful_ending_downtrack_;         // if algorithm was successful, this is traffic_light_downtrack
-    double last_successful_scheduled_entry_time_;     // if algorithm was successful, this is also scheduled entry time (ET in TSMO UC2 Algo)
-
-    carma_planning_msgs::msg::Plugin plugin_discovery_msg_;
-    carma_debug_ros2_msgs::msg::TrajectoryCurvatureSpeeds debug_msg_;
-    std::vector<double> last_final_speeds_;
-
-    std::string light_controlled_intersection_strategy_ = "signalized"; // Strategy carma-streets is sending. Could be more verbose but needs to be changed on both ends
+    std::shared_ptr<LightControlledIntersectionTacticalPlugin> worker_;
 
   public:
   
@@ -145,60 +62,6 @@ namespace light_controlled_intersection_tactical_plugin
      */
     rcl_interfaces::msg::SetParametersResult 
     parameter_update_callback(const std::vector<rclcpp::Parameter> &parameters);
-
-    /**
-     * \brief Creates a speed profile according to case one or two of the light controlled intersection, where the vehicle accelerates (then cruises if needed) and decelerates into the intersection. 
-     * \param wm world_model pointer
-     * \param points_and_target_speeds of centerline points paired with speed limits whose speeds are to be modified:
-     * \param start_dist starting downtrack of the maneuver to be planned (excluding buffer points) in m
-     * \param remaining_dist distance for the maneuver to be planned (excluding buffer points) in m
-     * \param starting_speed starting speed at the start of the maneuver in m/s
-     * \param departure_speed ending speed of the maneuver a.k.a entry speed into the intersection m/s
-     * \param tsp trajectory smoothing parameters
-     * NOTE: Cruising speed profile is applied (case 1) if speed before deceleration is higher than speed limit. Otherwise Case 2.
-     * NOTE: when applying the speed profile, the function ignores buffer points beyond start_dist and end_dist. Internally uses: config_.back_distance and speed_limit_
-     */
-    void apply_trajectory_smoothing_algorithm(const carma_wm::WorldModelConstPtr& wm, std::vector<PointSpeedPair>& points_and_target_speeds, double start_dist, double remaining_dist, 
-                                            double starting_speed, double departure_speed, TrajectoryParams tsp);
-
-    /**
-     * \brief Apply optimized target speeds to the trajectory determined for fixed-time and actuated signals.
-     *        Based on TSMO USE CASE 2. Chapter 2. Trajectory Smoothing
-     * \param maneuver Maneuver associated that has starting downtrack and desired entry time
-     * \param starting_speed Starting speed of the vehicle
-     * \param points The set of points with raw speed limits whose speed profile to be changed.
-     *               These points must be in the same lane as the vehicle and must extend in front of it though it is fine if they also extend behind it. 
-     *              
-     */
-    void apply_optimized_target_speed_profile(const carma_planning_msgs::msg::Maneuver& maneuver, const double starting_speed, std::vector<PointSpeedPair>& points_and_target_speeds);
-
-    /**
-     * \brief Creates geometry profile to return a point speed pair struct for INTERSECTION_TRANSIT maneuver types (by converting it to LANE_FOLLOW)
-     * \param maneuvers The list of maneuvers to convert to geometry points and calculate associated raw speed limits
-     *  \param max_starting_downtrack The maximum downtrack that is allowed for the first maneuver. This should be set to the vehicle position or earlier.
-     *                               If the first maneuver exceeds this then it's downtrack will be shifted to this value. 
-     * \param wm Pointer to intialized world model for semantic map access
-     * \param ending_state_before_buffer reference to Vehicle state, which is state before applying extra points for curvature calculation that are removed later
-     * \param state The vehicle state at the time the function is called
-     * \param general_config Basic autonomy struct defined to load general config parameters from tactical plugins
-     * \param detailed_config Basic autonomy struct defined to load detailed config parameters from tactical plugins
-     * \return A vector of point speed pair struct which contains geometry points as basicpoint::lanelet2d and speed as a double for the maneuver
-     * NOTE: This function is a slightly modified version of the same function in basic_autonomy library and currently only plans for the first maneuver
-     */
-    std::vector<PointSpeedPair> create_geometry_profile(const std::vector<carma_planning_msgs::msg::Maneuver> &maneuvers, double max_starting_downtrack,const carma_wm::WorldModelConstPtr &wm,
-                                                                        carma_planning_msgs::msg::VehicleState &ending_state_before_buffer,const carma_planning_msgs::msg::VehicleState& state,
-                                                                        const GeneralTrajConfig &general_config, const DetailedTrajConfig &detailed_config);
-
-    /**
-     * \brief Given a Lanelet, find it's associated Speed Limit
-     *
-     * \param llt Constant Lanelet object
-     *
-     * \throw std::invalid_argument if the speed limit could not be retrieved
-     *
-     * \return value of speed limit in mps
-     */
-    double findSpeedLimit(const lanelet::ConstLanelet& llt, const carma_wm::WorldModelConstPtr &wm) const;
 
     ////
     // Overrides
